Edit Post: Refactor effects to action generators (#27069)
* Edit Post: Refactor effects to action generators

* Edit Post: Remove effects middleware and use register()

* Edit Post: Remove refx and add @wordpress/data-controls

* Edit Post: Delete unused effects and middlewares

* Edit Post: Fix tests

* Edit Post: Revert store registration

* Use global saveMetaboxUnsubscribe

* Enable the data controls in the edit-post registry

* Do not wait for requestMetaBoxUpdates to finish
